Extracted defaults into a separate file
